CPP and TXT files for Linked List
This is a CPP file consisting of code that loads course data from the text file and creates a linked list of the course objects. Course objects consist of a string course name, string course number, and vector course prerequisites. The program handles input from the user via a menu, taking action based on the user's input to either load the data from the file, print all course data (using a insertion sort algorithm to print in alphanumeric order), print details of a specific course (using a function that searches the linked list with the passed string course number parameter), or exits the menu.